refactor: restructure bookmarks
