+++ src/App.tsx
@@ -5,7 +5,7 @@ import { createStackNavigator } from '@react-navigation/stack'
 import withContexts from './components/with-contexts'
 import { searchScreen } from './screens/Search'
 import { gurbaniScreen } from './screens/Gurbani'
-import { bookmarksScreen } from './screens/BookmarksScreen'
+import { bookmarksScreen } from './screens/Bookmarks'
 import { AppStackParams } from './lib/screens'
 
 const screens = [ gurbaniScreen, searchScreen, bookmarksScreen ]
